remove unsafe function, replace with v2
The old API function `vi_to_uint64()` was deemed unsafe some time ago as
it provides no way to detect or prevent reading over the end of a buffer.
The newer `vi_to_uint64_2()` was developed with an additional argument to
supply the input buffer length, and error checking returning EINVAL in
such case as well as ERANGE when VI_MAX would be surpassed.

This change removes the old function, renames the new function to match
the old, updates the reference in calls to the new function, and provides
the correct new buffer length argument to calls of the old function.